refactor: 날씨 관련 클래스 패키지 구조 재구성
