+++ src/main/java/com/backendoori/ootw/post/dto/PostSaveResponse.java
@@ -2,6 +2,7 @@ package com.backendoori.ootw.post.dto;
 
 import java.time.LocalDateTime;
 import com.backendoori.ootw.post.domain.Post;
+import com.backendoori.ootw.weather.dto.WeatherDto;
 
 public record PostSaveResponse(
     Long postId,
