+++ src/main/java/com/backendoori/ootw/weather/domain/forecast/BaseDateTime.java
@@ -1,4 +1,4 @@
-package com.backendoori.ootw.weather.domain;
+package com.backendoori.ootw.weather.domain.forecast;
 
 public record BaseDateTime(
     String baseDate,
