+++ src/main/java/com/backendoori/ootw/weather/service/WeatherService.java
@@ -3,8 +3,8 @@ package com.backendoori.ootw.weather.service;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
-import com.backendoori.ootw.weather.domain.BaseDateTime;
-import com.backendoori.ootw.weather.domain.ForecastCategory;
+import com.backendoori.ootw.weather.domain.forecast.BaseDateTime;
+import com.backendoori.ootw.weather.domain.forecast.ForecastCategory;
 import com.backendoori.ootw.weather.dto.WeatherResponse;
 import com.backendoori.ootw.weather.util.BaseDateTimeCalculator;
 import com.backendoori.ootw.weather.util.client.ForecastApiClient;
