+++ src/test/java/com/backendoori/ootw/post/controller/PostControllerTest.java
@@ -15,12 +15,12 @@ import com.backendoori.ootw.exception.ExceptionResponse.FieldErrorDetail;
 import com.backendoori.ootw.post.dto.PostReadResponse;
 import com.backendoori.ootw.post.dto.PostSaveRequest;
 import com.backendoori.ootw.post.dto.PostSaveResponse;
-import com.backendoori.ootw.post.dto.WeatherDto;
 import com.backendoori.ootw.post.repository.PostRepository;
 import com.backendoori.ootw.post.service.PostService;
 import com.backendoori.ootw.security.TokenMockMvcTest;
 import com.backendoori.ootw.user.domain.User;
 import com.backendoori.ootw.user.repository.UserRepository;
+import com.backendoori.ootw.weather.dto.WeatherDto;
 import net.datafaker.Faker;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
