+++ src/test/java/com/backendoori/ootw/post/service/PostServiceTest.java
@@ -13,11 +13,11 @@ import com.backendoori.ootw.exception.UserNotFoundException;
 import com.backendoori.ootw.post.dto.PostReadResponse;
 import com.backendoori.ootw.post.dto.PostSaveRequest;
 import com.backendoori.ootw.post.dto.PostSaveResponse;
-import com.backendoori.ootw.post.dto.WeatherDto;
 import com.backendoori.ootw.post.dto.WriterDto;
 import com.backendoori.ootw.post.repository.PostRepository;
 import com.backendoori.ootw.user.domain.User;
 import com.backendoori.ootw.user.repository.UserRepository;
+import com.backendoori.ootw.weather.dto.WeatherDto;
 import net.datafaker.Faker;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.AfterAll;
